Revert "Fix multi download screen"
