+++ app/src/main/AndroidManifest.xml
@@ -34,9 +34,7 @@
             </intent-filter>
         </activity>
 
-        <activity
-            android:launchMode= "singleTop"
-            android:name=".feature_downloader.presentation.DownloadManagementActivity" />
+        <activity android:name=".feature_downloader.presentation.DownloadManagementActivity" />
 
         <activity
             android:name="com.google.android.gms.oss.licenses.OssLicensesMenuActivity"
