+++ app/src/main/res/values/strings.xml
@@ -2,5 +2,4 @@
     <string name="app_name">AudioBook</string>
     <string name="title_feature_book">Feature Book</string>
     <string name="title_feature_downloader">Download Feature</string>
-    <string name="tooltip_open_player_message">Swipe up to open player</string>
 </resources>
