+++ common/src/main/java/com/allsoftdroid/common/base/store/audioPlayer/AudioPlayerEvent.kt
@@ -13,7 +13,6 @@ data class Pause(val result: AudioPlayerEventState) : AudioPlayerEvent()
 data class EmptyEvent(val default: AudioPlayerEventState):AudioPlayerEvent()
 object Rewind : AudioPlayerEvent()
 object Forward : AudioPlayerEvent()
-object Finished : AudioPlayerEvent()
 
 //Details or information event for the player and UI
 data class PlaySelectedTrack(val trackList : List<AudioPlayListItem>,val bookId:String,val bookName:String, val position:Int) : AudioPlayerEvent()
