+++ database/src/main/java/com/allsoftdroid/database/bookListDB/DatabaseEntities.kt
@@ -21,10 +21,7 @@ data class DatabaseAudioBook(
     var creator : String?,
 
     @ColumnInfo(name = "published_date")
-    var date: String?,
-
-    @ColumnInfo(name = "added_date")
-    var addeddate: String?
+    var date: String?
 ){
-    constructor() : this("","","","","")
+    constructor() : this("","","","")
 }
