+++ feature_book/src/main/java/com/allsoftdroid/feature_book/domain/model/AudioBookDomainModel.kt
@@ -4,6 +4,5 @@ data class AudioBookDomainModel(
     val mId: String,
     val title: String,
     val creator: String?,
-    val date: String?,
-    val addeddate:String
+    val date: String?
 )
