+++ feature_book_details/src/main/java/com/allsoftdroid/feature/book_details/presentation/AudioBookDetailsFragment.kt
@@ -36,7 +36,6 @@ import org.koin.androidx.viewmodel.ext.android.viewModel
 import org.koin.core.KoinComponent
 import org.koin.core.parameter.parametersOf
 import timber.log.Timber
-import java.util.*
 
 
 class AudioBookDetailsFragment : BaseUIFragment(),KoinComponent {
@@ -295,29 +294,15 @@ class AudioBookDetailsFragment : BaseUIFragment(),KoinComponent {
 
         dataBinding.bookMediaActionsRead.setOnClickListener {
             bookDetailsViewModel.additionalBookDetails.value?.let {
-                if(it.gutenbergUrl.isNotEmpty() && isLinkValid(it.gutenbergUrl)){
+                if (it.archiveUrl.isEmpty()){
+                    Toast.makeText(this.requireActivity(),getString(R.string.no_link_found),Toast.LENGTH_SHORT).show()
+                }else{
                     val uri  = Uri.parse(it.gutenbergUrl)
                     val intent = Intent(Intent.ACTION_VIEW,uri)
                     startActivity(Intent.createChooser(intent,getString(R.string.open_with_label)))
-                }else if (it.archiveUrl.isNotEmpty() && isLinkValid(it.archiveUrl)){
-                    val uri  = Uri.parse(it.archiveUrl)
-                    val intent = Intent(Intent.ACTION_VIEW,uri)
-                    startActivity(Intent.createChooser(intent,getString(R.string.open_with_label)))
-                }else{
-                    Toast.makeText(this.requireActivity(),getString(R.string.no_link_found),Toast.LENGTH_SHORT).show()
                 }
             }?:Toast.makeText(this.requireActivity(),getString(R.string.wait_message),Toast.LENGTH_SHORT).show()
         }
-    }
-
-    private fun isLinkValid(archiveUrl: String): Boolean {
-        if (archiveUrl.toLowerCase(Locale.ROOT).contains("archive.org"))
-            return true
-
-        if (archiveUrl.toLowerCase(Locale.ROOT).contains("gutenberg.org"))
-            return true
-
-        return false
     }
 
     private fun resumeOrPlayFromStart() {
@@ -382,12 +367,6 @@ class AudioBookDetailsFragment : BaseUIFragment(),KoinComponent {
                     is EmptyEvent -> {
                         Timber.d("Initial event received")
                     }
-
-                    is Finished -> {
-                        Timber.d("Finished event received: Resetting list")
-                        bookDetailsViewModel.resetUI()
-                    }
-
                     else -> Timber.d("Unknown Pressed Details Fragment")
                 }
             }
