+++ feature_mini_player/src/main/java/com/allsoftdroid/audiobook/feature_mini_player/presentation/MiniPlayerFragment.kt
@@ -11,6 +11,7 @@ import com.allsoftdroid.audiobook.feature_mini_player.databinding.FragmentMiniPl
 import com.allsoftdroid.audiobook.feature_mini_player.di.FeatureMiniPlayerModule
 import com.allsoftdroid.audiobook.feature_mini_player.presentation.viewModel.MiniPlayerViewModel
 import com.allsoftdroid.common.base.fragment.BaseContainerFragment
+import it.sephiroth.android.library.xtooltip.Tooltip
 import org.koin.android.ext.android.inject
 import timber.log.Timber
 
@@ -20,6 +21,7 @@ class MiniPlayerFragment : BaseContainerFragment() {
     Lazily initialize the view model
      */
     private val miniPlayerViewModel: MiniPlayerViewModel by inject()
+    var tooltip:Tooltip? = null
     private lateinit var refBinding:FragmentMiniPlayerBinding
 
     override fun onCreateView(
@@ -44,5 +46,28 @@ class MiniPlayerFragment : BaseContainerFragment() {
         refBinding = binding
 
         return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        val metrics = resources.displayMetrics
+        val gravity = Tooltip.Gravity.TOP
+
+        tooltip = Tooltip.Builder(view.context)
+            .anchor(refBinding.ivMiniPlayerAlbumArt,70,70,false)
+            .text(getString(R.string.tooltip_open_player_message))
+            .maxWidth(metrics.widthPixels / 2)
+            .arrow(true)
+            .floatingAnimation(Tooltip.Animation.DEFAULT)
+            .showDuration(3000)
+            .overlay(true)
+            .create()
+
+        tooltip
+            ?.doOnHidden {
+                tooltip = null
+            }
+            ?.show(refBinding.root, gravity, true)
     }
 }
