+++ feature_mini_player/src/main/res/layout/fragment_mini_player.xml
@@ -22,6 +22,7 @@
             android:layout_marginBottom="8dp"
             android:contentDescription="@string/audio_book_album_art"
             android:padding="@dimen/mini_player_icon_padding"
+            android:onClick="@{() -> viewModel.openMainPlayer()}"
             app:bookImage="@{viewModel.bookId}"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintStart_toStartOf="parent"
