+++ feature_mini_player/src/test/java/com/allsoftdroid/audiobook/feature_mini_player/presentation/viewModel/MiniPlayerViewModelTest.kt
@@ -101,4 +101,21 @@ class MiniPlayerViewModelTest{
             }
         }
     }
+
+    @Test
+    fun openMainPlayer_sentEvent_returnNothing(){
+        playerViewModel.openMainPlayer()
+
+        userActionEventStore.observe().subscribe {
+            it.getContentIfNotHandled()?.let {event->
+
+                val eventType = when(event){
+                    is OpenMainPlayerUI -> "MainPlayer"
+                    else -> "Not MainPlayer"
+                }
+                assertThat(eventType, `is`("MainPlayer"))
+            }
+        }
+    }
+
 }
