+++ feature_playerfullscreen/src/main/java/com/allsoftdroid/audiobook/feature/feature_playerfullscreen/presentation/MainPlayerFragment.kt
@@ -87,10 +87,6 @@ class MainPlayerFragment : BaseContainerFragment(){
                                 setImageResource(R.drawable.play_circle_green)
                             }
                         }
-
-                        is Finished -> {
-                            mainPlayerViewModel.bookFinished()
-                        }
                         else -> Timber.d("Ignore event $event")
                     }
                 }
@@ -127,7 +123,7 @@ class MainPlayerFragment : BaseContainerFragment(){
 
     private fun handleBackPressEvent(){
         this.findNavController().navigateUp()
-        mainPlayerViewModel.showMiniPlayerIfPlaying()
+        mainPlayerViewModel.showMiniPlayer()
     }
 
     override fun onDestroyView() {
