+++ services/src/main/java/com/allsoftdroid/audiobook/services/audio/utils/NotificationUtils.kt
@@ -108,7 +108,7 @@ class NotificationUtils {
             if (Build.VERSION.SDK_INT>= Build.VERSION_CODES.O) {
                 val notificationManager = applicationContext.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
                 val name = applicationContext.getString(R.string.app_name)
-                val importance = NotificationManager.IMPORTANCE_LOW
+                val importance = NotificationManager.IMPORTANCE_DEFAULT
                 NotificationChannel(NOTIFICATION_CHANNEL, name, importance).apply {
                     enableLights(false)
                     enableVibration(false)
