--- services/src/main/java/com/allsoftdroid/audiobook/services/audio/utils/PlayerState.kt
@@ -1,13 +1,0 @@
-package com.allsoftdroid.audiobook.services.audio.utils
-
-enum class PlayerState{
-    SourceError,
-    SystemError,
-
-    PlayerFinished,
-    PlayerReady,
-    PlayerBusy,
-    PlayerIdle,
-    PlayingNext
-}
-
